ca-certificates: fix on-target postinstall script
When the package is installed directly on the machine (instead of
installing it in the rootfs directly), the postinstall script fails with
the following error:

/usr/sbin/update-ca-certificates: line 75: shift: shift count out of range

The reason is that the "update-ca-certificates" script is executed with
the "--sysroot" argument, and as the sysroot $D is passed. However on the
target system this variable doesn't exist, so the argument is passed without
this mandatory value, and the execution fails.

To avoid this error, check if the $D variable exists, and pass the --sysroot
argument only when it does.

Reported-by: WXbet <Wxbet@proton.me>
Signed-off-by: Gyorgy Sarvari <skandigraun@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Upstream-Hash: cf39461e97098a1b28693299677888ba7e8bfccf
Upstream-Url: https://git.openembedded.org/openembedded-core/commit/?id=cf39461e97098a1b28693299677888ba7e8bfccf
Upstream-GitHub: https://github.com/openembedded/openembedded-core/commit/cf39461e97098a1b28693299677888ba7e8bfccf